requirements.txt
alembic==1.6.5
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
migrate==0.3.7
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.1
pandas==1.3.0
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==2.0.1
